requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
chromadb
sentence-transformers
python-dotenv
pypdf
